handling or repo names

+++ dockerfiles/deps/base.sh
@@ -8,7 +8,13 @@ PANDOC_VERSION=${PANDOC_VERSION:-"3.1.9"}
 echo "GPU enabled: $GPU_ENABLED"
 
 dnf -y update && dnf -y upgrade
-dnf -y install epel-release
+
+if [ "$DOCKERFILE" != "ubi9.4" ]; then
+  dnf -y install epel-release
+  REPO_NAME="Rocky-Devel"
+else
+  REPO_NAME="ubi"
+fi
 dnf -y install poppler-utils xz-devel wget tar make which mailcap dnf-plugins-core compat-openssl11
 
 ARCH=$(uname -m)
@@ -28,8 +34,10 @@ else
 fi
 
 # This is a fix for an bug where config-manager tries to modify a repo file with the incorrect name
-cp /etc/yum.repos.d/rocky-devel.repo /etc/yum.repos.d/Rocky-Devel.repo
-dnf config-manager --enable crb
+if [ "$DOCKERFILE" != "ubi9.4" ]; then
+  cp /etc/yum.repos.d/rocky-devel.repo /etc/yum.repos.d/"$REPO_NAME".repo
+  dnf config-manager --enable crb
+fi
 
 # Needed for LibreOffice to install base components on aarch64
 if [ "$ARCH" == "aarch64" ]; then
@@ -39,14 +47,19 @@ else
 fi
 pandoc_filename=pandoc-"$PANDOC_VERSION"-linux-"$PANDOC_ARCH".tar.gz
 pandoc_url=https://github.com/jgm/pandoc/releases/download/"$PANDOC_VERSION"/"$pandoc_filename"
-sed -i 's/enabled=0/enabled=1/g' /etc/yum.repos.d/Rocky-Devel.repo
+sed -i 's/enabled=0/enabled=1/g' /etc/yum.repos.d/"$REPO_NAME".repo
 wget "$pandoc_url"
 tar xvzf "$pandoc_filename" --strip-components 1 -C '/usr/local'
 rm -rf "$pandoc_filename"
-dnf -y install libreoffice-writer libreoffice-base libreoffice-impress libreoffice-draw libreoffice-math libreoffice-core
-sed -i 's/enabled=1/enabled=0/g' /etc/yum.repos.d/Rocky-Devel.repo
+if [ "$DOCKERFILE" != "ubi9.4" ]; then
+  dnf -y install libreoffice-writer libreoffice-base libreoffice-impress libreoffice-draw libreoffice-math libreoffice-core
+fi
 
-# General cleanup
-rm -rf /var/cache/yum/*
-rm -f /etc/yum.repos.d/Rocky-Devel.repo
-dnf clean all
+if [ "$DOCKERFILE" != "ubi9.4" ]; then
+  sed -i 's/enabled=1/enabled=0/g' /etc/yum.repos.d/"$REPO_NAME".repo
+  rm -f /etc/yum.repos.d/"$REPO_NAME".repo
+  # General cleanup
+  rm -rf /var/cache/yum/*
+  dnf clean all
+fi
+
